Themes: Handle case when keyguard is showing during theme change
After we recreate the status bar we need to check if the keyguard
was showing and if so make sure it is showing after the theme
change.  We also need to make sure the phone status bar is not
visible if the keyguard was showing, otherwise you see two status
bars.

Change-Id: Iba09e3e5c8cf52d49f292423a50034a3cf84d9c1
Signed-off-by: AudioGod <audiogod@sonic-developers.com>