Themes: Drop timestamp and add change request type
This patch ditches the hack where we keep track of the time a theme
change occurs.  Instead we keep track of the RequestType when a
theme change occurs.  This extra information is much more useful
to apps that want to handle theme changes, i.e. SystemUI.  If a
new theme config is equal to an old theme config, the app can look
at the request type and determine what action to take.

Change-Id: Ie1d56060069d6645bf3212f4a70739d491cd0731
Signed-off-by: AudioGod <audiogod@sonic-developers.com>