Themes: Change themeChange config value to 0x300000
This value is ActivityInfo.CONFIG_THEME_RESOURCE and
ActivityInfo.CONFIG_THEME_FONT ORed together.  Any future theme
config flags that are added should be ORed with this value so that
any app that declares themeChange as a configuration change it will
handle will not be restarted when only one theme change flag is
set.

Change-Id: I620ab62a54f97c20e03369adf01ba817fada51de
Signed-off-by: AudioGod <audiogod@sonic-developers.com>