SysUI: Invert notification backgrounds
Notifications for ambient display were enhanced to support fading
back into full color and saturation.  This change involved hiding
the notification background instead of inverting it along with the
rest of the notification view.  This doesn't play well with themes
that use a darker background and lighter text color.

This patch keeps the background visible and inverts it along with
the rest of the notification, which allows the text to be visible.

Change-Id: I8e7cb5b77be096c6a426751ad4c98de2529059d1
Signed-off-by: AudioGod <audiogod@sonic-developers.com>