Track current user id in ThemeService
Intent.ACTION_USER_SWITCHED is broadcasted on boot which causes the
ThemeService's reciever to fire off and request a theme change.
This patch tracks the current user ID, with the default being
UserHandle.USER_OWNER, and will only request the theme change when
the new user differs from the stored current user id.

Change-Id: I69a2d9d669a0605c44eabaf6bf541494370e1e09
REF: CYNGNOS-429
Signed-off-by: AudioGod <audiogod@sonic-developers.com>